+++ CHANGELOG.md
@@ -19,7 +19,7 @@ Prefix the change with one of these keywords:
 - Added: `ssr` and `toggleAt` props to `FooterSection` component
 - Deprecated: `showAt` and `hideAt` props of `FooterSection` component
 - Fixed: unwanted props bleeding through to HTML attributes when using `Icon` and `Spinner`
-- Added: `Switch` component!️
+- Added: `Pagination` component
 
 ## [0.30.0]
 
@@ -31,10 +31,10 @@ Prefix the change with one of these keywords:
 - Removed: `ViewerContainer` component
 - Added: Datepicker component
 - Changed: `Radio` and `Checkbox` component HTML div tag to span
+- Added: `Switch` component
 
 ## [0.29.3]
 
-- Added: Switch component
 - Fixed: issue with not programmically changing the active / initial tab on `Tabs` component
 - Deprecated: `initialTab` in `Tabs` component. Please use `activeTab` instead.
 - Added: Quote component